#2 Header and CPP split into interface and implementation (for the most part)
SIDE NOTE: decidedly not cleaner to see dependency chain, there's a lot of redundant includes, it's a bigger set of files to work with - BUT: it is following a best practice, implementation is separate and all can compile separately. I will keep the legacy dependency chain in EntryPoint.cpp for reference.

Release build and .zip updated